requirements.txt
# Core AI/ML Dependencies
langchain==0.2.16
langchain-openai==0.1.23
langgraph==0.2.16
langsmith==0.1.98

# API and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
httpx==0.27.0
aiohttp==3.10.5

# Data Processing and Analysis
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
statsmodels==0.14.2
prophet==1.1.5

# Visualization
matplotlib==3.9.2
plotly==5.22.0
seaborn==0.13.2

# GitHub API
PyGithub==2.3.0
requests==2.32.3

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
python-multipart==0.0.9
jinja2==3.1.4

# Development
pytest==8.3.2
black==24.8.0
isort==5.13.2

